Don't remove force quit script

+++ GenerateGamesList.py
@@ -22,7 +22,8 @@ def clear_directory(folder_path):
             file_path = os.path.join(folder_path, the_file)
             if os.path.isfile(file_path) \
                 and the_file != RefreshListScript \
-                and not the_file.endswith(('.txt', 'py')):
+                and not the_file.endswith(('.txt', 'py')) \
+                and not the_file == "Force_Quit.sh":
                     os.unlink(file_path)
         except Exception as e:
             print(e)
